+++ skeleton/Program.h
@@ -43,11 +43,11 @@ private:
 
 	std::vector<const Renderable*> objects;
 
-	float radius;
+	double radius;
 
-	float scale;
-	float latRot;
-	float longRot;
+	double scale;
+	double latRot;
+	double longRot;
 
 	void setupWindow();
 	void insertAirSigmets();
